Cleanup unittest and yolort.utils
