+++ yolort/utils/hooks.py
